requirements.txt
telethon
discord.py
requests
python-dotenv
apscheduler
openai
# Optional: anthropic